requirements.txt
requests==2.18.4
PyJWT==1.5.3
cryptography==2.0.3